flask-app/requirements.txt
flask>=3.0,<4.0
webexteamssdk>=1.6,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
pyngrok>=7.0,<8.0
